Delete DummyJSONUser type
