--- src/lib/types/DummyJSONUser.ts
@@ -1,9 +1,0 @@
-type DummyJSONUser = {
-  id: number;
-  firstName: string;
-  lastName: string;
-  username: string;
-  image: string;
-};
-
-export default DummyJSONUser;
